+++ packages/akashic-cli-serve/src/client/view/molecule/EntityTreeOptionBar.tsx
@@ -19,6 +19,7 @@ export class EntityTreeOptionBar extends React.Component<EntityTreeOptionBarProp
 		const { isSelectingEntity, showsHidden, onClickSelectEntity, onClickUpdateEntityTrees } = this.props;
 		return <div className={styles["entity-tree-option-bar"]}>
 			<ToolIconButton
+				className="external-ref_button_select-entity-from-screen"
 				icon="input"
 				title="ゲーム画面からエンティティを選択"
 				pushed={isSelectingEntity}
@@ -27,6 +28,7 @@ export class EntityTreeOptionBar extends React.Component<EntityTreeOptionBarProp
 			/>
 			<div className={styles["sep"]} />
 			<ToolIconButton
+				className="external-ref_button_refresh-entity-tree"
 				icon="refresh"
 				title="エンティティツリーを更新"
 				size={20}
@@ -39,6 +41,7 @@ export class EntityTreeOptionBar extends React.Component<EntityTreeOptionBarProp
 				Show hidden
 			</label>
 			<ToolIconButton
+				className="external-ref_button_dump-to-console"
 				icon="web_asset"
 				title="選択エンティティをコンソールにダンプ"
 				size={20}